requirements.txt
asgiref==3.7.2
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.9
gunicorn==20.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.1
filelock==3.13.1
httplib2==0.22.0
idna==3.6
platformdirs==3.11.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.24.6
